+++ package.json
@@ -35,6 +35,8 @@
     "@crowdin/crowdin-api-client": "^1.10.4",
     "@ethersproject/abi": "^5.0.7",
     "@pancakeswap-libs/uikit": "^0.10.5",
+    "@ethersproject/providers": "^5.0.19",
+    "@pancakeswap-libs/sdk": "^1.0.1",
     "@reduxjs/toolkit": "^1.5.0",
     "@types/react": "^17.0.2",
     "@types/react-dom": "^17.0.1",
