+++ src/connectors/RequestError.ts
@@ -1,0 +1,7 @@
+class RequestError extends Error {
+  constructor(message: string, public code: number, public data?: unknown) {
+    super(message)
+  }
+}
+
+export default RequestError
