+++ src/connectors/bsc/noBscProviderError.ts
@@ -1,0 +1,9 @@
+class NoBscProviderError extends Error {
+  public constructor() {
+    super()
+    this.name = this.constructor.name
+    this.message = 'No BSC provider was found on window.BinanceChain.'
+  }
+}
+
+export default NoBscProviderError
