+++ src/connectors/bsc/userRejectedRequestError.ts
@@ -1,0 +1,9 @@
+class UserRejectedRequestError extends Error {
+  public constructor() {
+    super()
+    this.name = this.constructor.name
+    this.message = 'The user rejected the request.'
+  }
+}
+
+export default UserRejectedRequestError
